Get destination cell from directed edge

Get an H3 index representing the destination of a directed edge.

get_uddest(h3_edge = NULL, simple = TRUE)

Arguments

h3_edge

Address of directed edge.

simple

Logical; whether to return a vector of outputs or a data frame containing both inputs and outputs.

Value

By default, character vector of h3 cell indexes.

Examples

# Get the destination cell index of this directed edge index
get_uddest(h3_edge = '166be8d12fffffff')
#> [1] "86be8d127ffffff"